Add function to display status of the match
Adds a new function which displays match status, such as the current
turn, and displays the results of the match when the game is over. New
logic is added to the gridClickHandler function as well which helps move
the state of the game along as the player makes their attack choices.
Finally the eventListener used to record player attacks is moved from
the grid to the individual square div tiles to fix an issue where
clicking on the same square would select the entire board.