Add a function which places all of comp's ships
Adds a function to randomly place all 5 ships on the opponent's(comp's)
board. This function replaces some of placeShips' functionality and so
introduces a breaking change. Many of the test cases that used or tested
placeShips have been tweaked and an additional test case has been added
for the new function.