Add functionality to allow choice ship placement
Adds new function to the GameModule module which allows the player to
choose where to place each ship on their own board. This replaces the
old placeShips function and so many test cases are adjusted to avoid
conflicts with the new change. An appropriate test case is added for the
new function.