Add a prompt to get ship placements from player
Adds a function which uses prompts to get the placement coordinates for
each of the ships from the player. Helper functions for this are also
added and a function to turn on eventListeners for the opponent's board
is introduced to prevent attacks from being made before player has setup
their board.