Remove playerTurn variable entirely from GameModule
Updates GameModule factory function by removing the playerTurn variable
entirely. This variable was intended to track the player's, or
opponent's, turn and thus keep track of which player should be allowed
to attack, but ended up not being utilized. The test cases related to
this functionality have been removed as well.