Fix logic issues with promptShipPlacements
Updates promptShipPlacements with new GameModule functions to make the
function more succinct. Also fixes issues with logic during the while
loop that takes ship coordinates where it was possible to setup an
incomplete board and also issues where user could be stuck in a prompt
loop. New function, startGame, is introduced which allows user attacks
on the enemy board provided that the player board was setup fully
beforehand.